+++ test/test_scan_apache.rb
@@ -26,6 +26,29 @@ class TestScannerApache < Minitest::Test
     server.exit
   end
 
+  def test_check_tomcat_2019_0232
+    override_stdout
+
+    port = rand(60000) + 1024 # pick a random port number
+    server = start_web_server File.dirname(__FILE__) + '/data/apache_server_info.txt', '/cgi-bin/test.bat', port
+    uri = URI.parse "http://localhost:#{port}/cgi-bin/test.bat"
+    links = [uri.to_s]
+
+    error = nil
+    begin
+      Yawast::Scanner::Plugins::Servers::Apache.check_cve_2019_0232 links
+    rescue => e
+      error = e.message
+    end
+
+    assert !stdout_value.include?('[V]'), "Unexpected finding: #{stdout_value}"
+    assert error == nil, "Unexpected error: #{error}"
+
+    restore_stdout
+
+    server.exit
+  end
+
   def test_check_struts2_samples
     override_stdout